EZPixelDumperX2/pyproject.toml
[project]
name = "pixeldumper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dxcam==0.0.5",
    "flask==3.1.2",
    "numpy==2.4.2",
    "opencv-python==4.13.0.90",
    "pillow==12.1.1",
    "pyside6==6.10.1",
    "pywin32==311",
    "xxhash==3.6.0",
]
